fix(server): partition clustering test suites by behavior area
Qodo round-2 follow-up on rule 513872: the split-out test files were
still monolithic. Shared fixtures and stubs stay in each tests.rs;
test functions move to responsibility-focused child modules —
ordered_relay/tests/{sender,receiver,muc_validation}.rs,
relay/tests/{error_classification,dispatch,dispatch_span}.rs,
route_bridge/tests/{delivery,nack_channels}.rs. The dispatch-span
guard's include_str path follows the move; no test bodies changed.
